Rework on topics page
