+++ src/components/HomePageProducer.vue
@@ -71,7 +71,14 @@ export default {
     onEdit(targetKey, action) {
       console.log(targetKey, action);
       if (action === "remove") {
+        let toRemovePanName = this.panes.find(
+          (pane) => pane.key === targetKey
+        ).title;
+        console.log(toRemovePanName);
         this.panes = this.panes.filter((pane) => pane.key !== targetKey);
+        this.addedPanes = this.addedPanes.filter(
+          (paneName) => paneName !== toRemovePanName
+        );
       }
     },
     handletopicselection(val) {
